Moved the subclasses of user to the same python file for easier access to the classes and methods
user.whatever.whatever rather than
user.whatever.whatever, admin.whatever.whatever, host.whatever.whatever, atendee.whatever.whatever
