--- PassIO/admin.py
@@ -1,15 +1,0 @@
-import venue, event
-
-class Admin:
-
-    def __init__(self, name, email, password, admin_key):
-        super().__init__(name, email, password)
-        self.__admin_key = admin_key
-
-    def __verify_venue(venue):
-        # TODO: complete function
-        return True
-
-    def __verify_event(event):
-        # TODO: complete function
-        return True
